opusenc_sys

Function opus_decoder_ctl

pub unsafe extern "C" fn opus_decoder_ctl(
    st: *mut OpusDecoder,
    request: c_int,
    ...
) -> c_int

Perform a CTL function on an Opus decoder.

Generally the request and subsequent arguments are generated by a convenience macro. @param st OpusDecoder*: Decoder state. @param request This and all remaining parameters should be replaced by one of the convenience macros in @ref opus_genericctls or @ref opus_decoderctls. @see opus_genericctls @see opus_decoderctls