opusenc_sys

Function opus_encode

pub unsafe extern "C" fn opus_encode(
    st: *mut OpusEncoder,
    pcm: *const opus_int16,
    frame_size: c_int,
    data: *mut c_uchar,
    max_data_bytes: opus_int32,
) -> opus_int32

Encodes an Opus frame. @param [in] st OpusEncoder*: Encoder state @param [in] pcm opus_int16*: Input signal (interleaved if 2 channels). length is frame_sizechannelssizeof(opus_int16) @param [in] frame_size int: Number of samples per channel in the input signal. This must be an Opus frame size for the encoder’s sampling rate. For example, at 48 kHz the permitted values are 120, 240, 480, 960, 1920, and 2880. Passing in a duration of less than 10 ms (480 samples at 48 kHz) will prevent the encoder from using the LPC or hybrid modes. @param [out] data unsigned char*: Output payload. This must contain storage for at least \a max_data_bytes. @param [in] max_data_bytes opus_int32: Size of the allocated memory for the output payload. This may be used to impose an upper limit on the instant bitrate, but should not be used as the only bitrate control. Use #OPUS_SET_BITRATE to control the bitrate. @returns The length of the encoded packet (in bytes) on success or a negative error code (see @ref opus_errorcodes) on failure.